Separate module tests from load tests
LOAD does a lot of things, so it's not clear what category it's in
(scanner? misc? system? series?)

But there's certainly plenty to be tested about modules without them
being lumped in with load tests.  In fact, they might need to have
the usermode MODULE command split out from tests of the MODULE!
datatype itself.