Align PROTECT with being Var-based, not Key-based
R3-Alpha put protection bits on keys, which were full cells and had
the bits to spare.

But since keylists could be shared this was bad.  Also, Ren-C now
compresses keys down to single pointers--not cells--there's no space
to store it in anyway.

This changes Protect_Key() to Protect_Var and makes the parameterization
more clear, in a way that works with Sea of Words.