Make Init_Any_Word_Bound() take symbol again
When trying to find ways to squeeze extra bits out of ANY-WORD! cells
in virtual binding, it was tried to say that bound words didn't store
a label, but would use their index to find the label.  This would
free up an entire pointer's worth of information.

This created trouble for case-insensitive behavior, as there was
nowhere to put the case variation.  So binding became case-sensitive
as a consequence.

The complexity related to this became too much, so it was switched
back to where symbols were stored in the words.  However the
functions for intializing bound words didn't allow passing in a
symbol...the index supplied was what determined the spelling.

However, modules now use INDEX_ATTACHED to bind words.  This meant
the same core routine could not be used for initializing bound
words, and so they were doing it in ad-hoc ways.

This puts things back to where the symbol is passed in.