Do not BIND/SET in Console Code Binding
The original Sea of Words implementation was ambitious in trying to
catch typos in SET-WORD! assignments.  This would mean needing a
new kind of primitive to transition SET-WORD! that were merely
attached into ones that were bound to actual variables.

To get the console working, it was changed to BIND/SET.  However that
turned all SET-WORD!s into ~unset~ in the user context in code
that was used.  So code like:

    >> pick [do: print:] 1
    == do:

...that would make DO and PRINT unset in the user context, overriding
their inherited definition.  That's clearly not tenable.

For the near term to get Sea of Words off the ground, it is allowing
all assignments of attached words to emerge them.  Design for a
"strict mode" adds many more usage issues, so better to start small.