Load make-file in user context in %bootstrap-shim.r
MAKE-FILE is included in modern Ren-C binaries, but older Ren-C doesn't
have it.  The script is in the %scripts/ directory, but it's a module
written to new (post-bootstrap executable) baseline.  However,
%bootstrap-shim.r definitions are only in the user context...leaving
LIB alone so as not to crash the older executable running its own
mezzanine.  This does a workaround to just DO the code and skip the
header saying it's a module.