Get System Functions by Symbol, not index
With Sea of Words, variables in modules are found by looking to the
global hash table of words for a symbol and following a linked list
to see if a "patch" holding a variable exists for that word.  Since
SYS is a module, we can use that facility.

Previously there was a scanning process that had to count the keys
in the sys context and number them, which now just adds complexity
and an additional %tmp-sysctx.h file.  That can be removed.