Add new answer for binding reflectors of "attached" words
The concept of a word being "attached" to a module--but no variable
backing it--is a new one.  It is not clear what the most useful answer
to BINDING OF is in such a situation, but NULL is misleading...

This makes the answer the BAD-WORD! of ~attached~.  It then adds a
new reflector ATTACH OF which will tell you the context to which it
is attached.  This avoids giving misleading information, but it will
take some usage and experience to know what is useful in practice.