Make attachment while scanning honor / and .
The anomalous PATH! and TUPLE! definitions that can act as WORD! were
not being bound by the implicit attachment phase in scanning.  The way
path elements were bound was that the individual words were bound and
then they were extracted, but the / and . are created out of blanks,
not out of WORD!s.

Solution is not very elegant but works for the / and . in Lib_Context.