Do not use BIND/SET in Interning, DO, IMPORT
The historical behavior of modules would do a scan over the contents
in advance--much as MAKE OBJECT! did--to know how many keys to make
for the module.  It looked for top-level set words.

This precluded constructions like:

     func: func [spec body] [...]

The reason being that the scan that picked up FUNC would unset it.

In Sea of Words it is more natural to allow the words to hold their
inherited values until they are explicitly overwritten.  This may
not be ideal for all situations, but for things like the bootstrap
shim it's important...because it becomes convoluted to have to hide
those declarations somewhere besides the top level.