Add %import-shim.r for bootstrapping
The R3-Alpha module system mechanisms are dfifferent from what is
being done in Ren-C "Sea of Words".  To write code that is compatible
between both, the path of least resistance is to essentially make the
bootstrap process load everything into the user context and avoid
trying to have the various rebmake and prep processes act as if
they are modularized.

However, new "Sea of Words" executables don't have a choice... as
right now DO of files will run them in isolated contexts and only
return one result value.  To avoid stylizing all bootstrap files
as wrapped in MAKE OBJECT!, this introduces one %import-shim.r that
you DO.  In a modern Ren-C this will act as a no-op and has no
need to export anything, because the existing IMPORT and EXPORT
are good enough.  But in the bootstrap executable it will give fake
IMPORT and EXPORT functions that mimic the interface for calling
but actually just put everything in the user context.