Fix memory leak of PG_Inaccessible_Varlist
Because varlists have SERIES_FLAG_DYNAMIC set in SERIES_MASK_VARLIST,
all allocations will have a secondary allocation of the array content.
But if SERIES_FLAG_INACCESSIBLE is set, then it is assumed that
content has already been freed.

This wasn't the case for PG_Inaccessible_Varlist, which just set
the flag without freeing the content.  This makes it a decayed form
of a singular to start with to sidestep the problem.