Temporarily disable process ID API in POSIX
A partially designed feature was added by Atronix to annotate
natives in extensions for which OS'es they were implemented for.
It's a reasonable idea, but it made the extension processing more
complex and was impeding progress.

Temporarily move the POSIX process ID aside while the module features
are being rewritten.