Further specify CSCAPE string interpolation bindings
The CSCAPE operation was an attempt to use string interpolation before
any real mechanisms to support it existed.  This means that it kind
of guessed at bindings and then you could give it more hints by
passing in things that might be in binding groups.

It's pretty bad, but the advantages in being able to read the
string expressions being escaped into were worth it.  You could fumble
with passing some words and see if it provided enough contexts to
find the named strings.

The "Sea of Words" module mechanism moves things around more, so the
lists had to be redone.  But really, this points to a hole in binding
that ideally could be fixed...to give strings enough contextual
information to be able to be interpolated.

That will have to wait for another day, but this just adds and
shuffles the example words passed in to make the Sea of Words version
find all the variables named to be substituted in the strings.