Redesign how NATIVEs are loaded (Boot+Extensions)
In R3-Alpha, NATIVE was a hand-made function that was created at
the start of boot.  The definitions of the other natives were actually
made by running that NATIVE function through the evaluator.

Some technical reasons made Ren-C need to change this somewhat early
on, due to the evaluator using native identities for various
purposes...which made it impossible to run the evaluator without
the natives being loaded.  Considering that the NATIVE native had
to be built by hand anyway, it seemed that building all the natives
by hand made more sense and was less convoluted.

Several things have changed since that time.  Firstly, the technical
limitation is gone--the evaluator does not need any particular natives
loaded in order to run.

Secondly, if NATIVE is an ordinary function then it can be folded in
with the source code for an extension at the top of the extension's
script initialization file...which has value because it means there
isn't a need for some out-of-band mechanism for getting the list of
natives and their loaded specs... they are loaded with the rest of
the source and registered by a call to NATIVE.

Thirdly, it became interesting over time to think of NATIVE as having
refinements.  One refinement example is /COMBINATOR, as most of the
code for registering a combinator is actually shared with registering
a non-combinator.

So this goes ahead and makes that change.  It means there is not a
separate "function specs list" passed to the extension collator--one
big script is loaded with the native registrations at the top, in the
order of the C function pointers passed to the collator.  Also, the
collator is given the original length in codepoints of the string--
not just the size in bytes.  This can be used to construct a TEXT!
which is known valid UTF-8, which may be able to be scanned faster
than if you didn't know it was already good.