Consolidate DO*, IMPORT*, and LOAD-MODULE
This eliminates redundant code and builds DO on top of IMPORT, so that
the common features are implemented in one place.  This includes:

* Changing the working directory to match the file path (or URL "path")
  of a script, and restoring the prior path on completion of failure

* Isolating the executed code into a MODULE! so that it doesn't leak
  into the caller's context.

* Turning <tag> identified script names to DO into a URL by looking it
  up in the modules library.  So `do <chess>` and `import <json>` use
  the same logic to figure out where that tag poitns.

* Setting system.script to reflect the executing script during its run,
  and then putting it back when control is returned to the caller.

But the actual action taken is different: DO returns the evaluative
product of the script as its primary result, while IMPORT returns the
module context the script was executed in.  And IMPORT is designed to
return the same context every time it is called--so modules are loaded
only once--while DO performs an action that you can run any number
of times.  So what DO does is effectively flips the order of the
return results of IMPORT.