Small tweaks to make bootstrap work
The %make-spec.reb files in extensions set up include files and
library settings for their compilation.  But they want to know
things like the repo-dir and output-dir.

Rather than making these config files aware of the internal details
of the %tools/ directory, this inserts the variables for the narrow
set of things they use into the spec for their creation.  The goal
is to keep it compiling while narrowing the complexity of that
"make-spec language".

Also, the code factored out for making the libRebol APIs for Javascript
and TCC was relying on being able to run in the same context as
the %make-reb-lib.r script, to have the APIs available as well as
all of the cscape/emit machinery.  It's a little tricky to think
of how to pass this in, so for now this just manually binds them
into the module that wraps make-reb-lib.r's execution.  Review.

With this change, bootstrap under Sea of Words works!