Add MAP, fix up FOR, reorganize loop oriented tests
This is a generalized set of upgrades to the looping constructs,
bringing MAP and FOR into parity as constructs that can work with EACH.

Updated tests serve as documentation for the changes.  The tests had
been in the %control/ category, but it seems better to put them in
their own category.

The names of the files had gotten quite out of sync with the functions,
which have been heavily reorganized in Ren-C.  This syncs them up
(while doing minimal editing on the messy files themselves).