Allow META-BLOCK! branches, Disable BLANK!/THE-GROUP!
This makes it legal to use META-BLOCK! branches, which are particularly
useful when coupled with MAP for achieving a "MAP-EACH/ONLY" type of
operation.

BLANK! branches raise some questions and have never been used.  Also,
with the @XXX forms under scrutiny when acting as branches, this
removes the @(...) branch construct just to minimize the spread of
the idea for now.