Fix off-by-one error in canon symbol table
The canon symbol table was not actually getting the last canon in
the array, it was short by one.

This wasn't noticed until a call to Canon(SYM_PARSE_REJECT) was tried.
That's the last symbol in the table (at this time), and it did not
work.  So this fixes the issue.