Use symbol table to find LIB definitions for natives
There was an array called "Natives" which stored the REBACT for natives
in the same index order as the C functions for natives in the array
built into the executable.

The purpose of this array was to give access to internal functions to
be able to run natives without going through symbol lookup in lib.
This was more optimal, but also protected any natives that made such
calls from being impacted by users overwriting lib definitions.

There was a negative effect of not using the function definitions in
lib, because the functions did not have their symbol cache which they
would have in a full value (it's not known from the REBACT* alone).

But also, the defensive reason doesn't make a lot of sense--because
if you tamper with LIB in a destructive way then the mezzanine would
break.  It's small comfort that natives-built-on-natives would still
be working (being able to swap implementations between native and
usermode code is something we expect to do anyway.)

The performance issue of looking up symbols being a little slower than
picking out of an array is something that another optimization is
looking to attack...to actually make a block of direct bindable
REBSER for immediately going from a symbol to a bindable stub in lib.
This would also double as the "canons table"...and give a fast non-hash
lookup that could find the "patches" table.  That optimization is
being designed, and would exceed the value of any optimization that
is perceived to be given by the fixed REBACT* table.