Create fixed size memory table for main LIB variables
This imposes an order on the SYM_XXX where the symbols for natives
and generics can be used to look up the variables by a fixed
array index from the symbol ID, without having to touch the canons
table or the global hash table for words.

It involved splitting %words.r into two files: one for which order
matters and needs to be folded in to the natives list in case there
is overlap in the terminology, and one in which order does not matter.
These files are %lib-words.r and %symbols.r