Load symbol strings from separate compressed series
The symbol strings which had SYM_XXX were being loaded as WORD!s by the
scanner.  This created a bit of a Catch-22 for things that wanted to
use the symbols during scanning...for instance to use the fake word
symbols for . and / when building them.  It also limited the ability
to use value cells living in LIB--such as those for BLANK or NULL.

Hacks were created to load specific symbol words prior to the scanning
of the boot block.  But this gets rid of the hacks--as well as making
the building of the table faster.  It separates out the symbol strings
into compressed bytes which are length-prefixed, hence they can be
interned without needing to calculate the length.

Speed is not the driving reason for doing it this way.  It just
simplifies things and helps get rid of hacks.