Convenience Lib() and Sys() access, consolidate globals
While writing functions in the core there was frequently need for
sample values, such as a blank or a null cell.  Some sample values
were constructed for this, and that involved overhead and explanation.

This seems redundant when it is considered that much of the system
will not function unless the Lib context contains NULL and BLANK in
their corresponding symbol slots.  However, historically the lib
context variables were in movable memory and could become expanded...
meaning it was costly to find those slots.

Now there is a portion of memory in fixed space that represents the
basic lib definitions.  This is just as fast to access as creating
global system variables at fixed addresses.

This consolidates those variables--as well as access to all the other
natives and features in that fixed block--with a macro called Lib().
A corresponding one called Sys() is also added for getting at
functions in the "sys context".