Remove quoting from "RunQ_Throws", merge into API
The first variadic routine for running functions was modeled after
a similar R3-Alpha function.  However it started using "feeds" and
went through the same type checking as everything else.

In order to avoid "double evaluation" of values, this went through
various workarounds that were ultimately resolved in the API as
"you have to quote it".  But this old routine was a holdout, that
enforced quoting mechanics by default.  It also was a holdout on
being converted to macros, as it was not exported in the API.

This dispenses with that and makes the functions leverage the
API machinery to get checking on variadics, implicit rebEND or use
of C++ parameter packs, etc.  It may be the functions should carry
an `_internal` label, but for now just name them like ordinary APIs.