Make bootstrap shim TRY defuse NULL fetch errors
The bootstrap executable triggers errors on the access of NULL words.
This behavior is not used any more now that bad-word isotopes are
used for error creation.  It creates friction in using NULL in the
bootstrap executable.

This makes the TRY operation use variadic trickery to look at the
expression that follows and process it in a way that will not trigger
an error on nulls, but will actually turn them to blanks.