COMMA! and reformatting of bootstrap scripts
There was a comment saying the bootstrap executable didn't support
commas.  However it does now through a shim (they're just stripped
out of non-string code).  This deletes that comment and went on a
light survey of ANY, ALL, and PARSE calls to throw in a few commas
where it made the code more legible