Add option for `===` divider to print as it goes by
This allows them to do double duty as comments and information in
higher verbosity runs of scripts.  The idea is in its infancy--and
at the moment if you turn it on you'll get it mixed up with
messages from the console.  So it needs work to separate out which
ones are applicable and which aren't (e.g. each module would define
its own instance).

The concept is being tried out in the whitespace interpreter, replacing
some of its --verbose > 0 PRINT strings.