Improve accuracy of Eval tick counting
For efficiency, R3-Alpha did not actually keep moment-to-moment track
of the total tick counts.  It instead would reconcile the total count
from a GC countdown timer that would count down to zero on each
evaluation.

However, this reconciliation was being thrown off...and there was no
"inefficient" count in the debug build to catch when it was incorrect.
This gave misleading reports of how many ticks had transpired...which
was making increase in memory use (setting signals often) incorrectly
say that there was an accompanying increase in ticks.

This fixes the reconciliation mechanism, and also adds a check in
the debug build that naively increments the counter in order to make
sure the optimized count is not off.