Speed up GET-BLOCK! branches, fix branch ACTION!s
The GET-BLOCK! form of branch had a bad interaction with predicates
in its use of REDUCE.  This is the canary in the coal mine for
quoted ACTION! predicates needing to go, so they are to be removed
in a future commit.

Regardless of that--this led to the observation that there's no need
to call REDUCE AS BLOCK! on the GET-BLOCK! in order to reduce it.
GET-BLOCK! itself runs in the evaluator as a reduce, so evaluate it.

(In today's world that costs the same as a REDUCE because it runs the
native, but it will be optimized.)

Another bug was happening in branches where END was being passed in
and passed as the argument to an action.  That was attempting to
quote the END.  This removes that attempt and just passes the END.