Unify %run-tests.r and %run-recover.r into one file
There's no real reason why one script could not take a command line
parameter to specify whether recovery was desired or not.  Having
two different entry points to running tests is confusing, and the
non-recovering form had atrophied.

This unifies the two.  Recovery is not currently working as it has
atrophied.  Generally this is not that important, as any crash in the
tests logs what test triggered the crash...and generally if it's fixed
you want to go back and run the tests from the top again.  But it's
a good feature to have, and should be tended to when time permits.