Fix caching problem with SET-WORD! and next gotten
Running functions flushes the f_next_gotten cache.  But a plain
assignment can cause trouble too, if it doesn't trigger a function
call:

     >> x: <before> x: 1 x
                         ^-- x value was cached din infix lookahead

It used to not be a problem, when variables didn't just pop into
existence or have their caching states changed.  But INDEX_ATTACHED
and the various complexities involved with that means we have to
flush on SET-WORD! if the symbols match.