Clean up test files for isolation
A new feature in the test suite allows each test to be run in isolation
unless they require to be run together (in a "cluster" indicated by
putting those tests in a block together).

This feature uncovered several cases where blocks were not actually
putting things in isolation with each other.  It also caught usages
of PRINT in the tests, and other errors.

It also pulls out the cases which had TEXT! inside of the BLOCK!
clusters, in case that has a better purpose than merely commenting.