Simplify test loading, run in isolated "cluster" modules
This takes out the custom parsing logic that tried to load individual
tests without getting a scanner failure.  One reason that feature
existed was to try and load tests that were written for varying
versions of the interpreter in the same file, e.g. Rebol2 vs. Rebol3,
and then be able to tag the tests as <R2> or <R3>.  This is an
unsolvable problem in a general sense...though it was interesting...
but it's not applicable right now and holds back development of more
interesting features.

One such feature is running tests in isolated module contexts.  This
does that, and simplifies the higher level logic in the process.