Remove attached status checking from HELP
At some point HELP was changed to check for attached vs. bound status
and then not looking up the word.  This makes no real sense because
basically every word in the user context is just "attached", so
HELP APPEND wouldn't work, etc.

It's not clear why that was added.  This takes it out and the HELP
tests still run.  Removed.