Make CONTINUE with no argument act as ~void~
The ability for BREAK to take an argument was removed due to a desire
to consistently know from the outside of a loop if it had been broken,
and reserving pure NULL as that signal.  However, CONTINUE was able
to take an argument--and the philosophy was that it would then act
just as if the body had run to completion and returned that value.

This created contention because there was no value that a no-argument
CONTINUE could take that would allow it to mean "don't add any value
to the MAP-EACH" and "don't break an UNTIL loop".  CONTINUE pretending
the loop body produced NULL could work for the former but not the
latter.

Today's isotopes permit the use of the ~void~ isotope as the requisite
intent; it is neither true nor false, and UNTIL can make a decision
that IF doesn't have to make that a void means "neither".  MAP-EACH
can group void isotopes in with NULL as another thing that means
"don't collect anything".