Add ~inherited~ answer for BINDING
The ~attached~ state as an answer for BINDING OF was meant to avoid
miscommunicating about the context of a variable, and letting people
get a finer-grained answer to it for modules.  Otherwise you would
have an empty user context and say `binding of 'append` and it would
tell you it was in the lib context...which would be quite a surprise
if you assigned it and found the lib context variable did not change.

However this failed to communicate the difference between whether there
was an inherited value or not.  So `binding of 'append` said ~attached~
and so did `binding of 'asdfasdfasdf`.  Knowing whether a get of the
variable could actually get anything is an important thing to know,
and it was necessary to write a function in HELP.

This adds ~inherited~ as the answer in BINDING OF when there is a
variable, but not in the context the binding is to.