Overhaul of handling of ~null~/~blank~/~false~ isotopes
An early attempt to smooth over the handling of ~null~ isotopes turned
out to take away from their usefulness...which is that a normal
parameter to a function would see them as regular NULL.  This weakened
the hand of isotopes, compared to asking the user to reconstitute the
value to NULL when they specifically knew it to be safe.

Additionally, this rethinks MATCH to actually give back distinct
~false~, ~blank~, and ~null~ isotopes when it matches a falsey input.
This way the user can reconstitute the intent via an operator (DECAY)
when it is appropriate to do so.  This removes the need for the MATCH*
form, as well as making it clear that NON is an "anti-ENSURE" and
not an "anti-MATCH" which doesn't make a lot of sense.