Avoid null isotope in common-parsers
There were two instances of:

    all [
        ...
        attempt [if condition [...]]
        ...
    ]

When the IF fails, it is NULL, which the ATTEMPT considers a case
where it's appropriate to make a ~null~ isotope...because there was
not actually code that failed.  (NULL is the distinct case reserved
to communicate failure.)

This changes it to have an ELSE branch which purposefully returns false.