Reinstate UPARSE's RETURN combinator
The RETURN keyword in PARSE has had a controversial history.  So it has
been added and dropped a few times.

One issue with it was that it would tamper with the expected return
type.  PARSE traditionally was narrowed to some "expected" result...
either LOGIC!, the input series, or how far the series got in the parse.
Having RETURN would widen this to ANY-VALUE! and confuse the contract.

(That issue is no longer relevant, as the parse returns the synthesized
result of the block rule it gets, which can be any-value!.)

A second issue was that usages which were trying to "grab and go" of
a value out of a parse without needing to name or set a variable could
wind up neglecting matching unfinished rules.  The potential for
making mistakes based on eagerness to use the convenience of RETURN
when it didn't fit seemed high.

(That issue is less relevant, because the seamless way in which the
UPARSE construct synthesizes results means the only reason you would
reach for the RETURN would be if abrupt termination was what you
actually were desiring.)

A third issue is that RETURN being a parse keyword could confuse it
with the RETURN of a function, and make it hard to tell whether you
were returning from the parse or the function it was in.

(This is just life in the parse dialect, words mean something different
and that's true of WHILE and ANY and everything else.)

It was always the case that RETURN was going to be something you could
create, just uncertain if it would be in the box of default combinators.
But it seems like it may be a clearer solution for people to get their
head around than having to learn about some primitive like `UPARSE*`
in order to get an "early result".  So it's back in.