Replace device Init and Quit w/extension Start/Shutdown
The "Device" model was an attempt to build a raw C extension mechanism
which was disjoint from knowledge of the Rebol layer.

This led to issues of fixed enum values, and arrays of function
pointers in fixed ordering.  The theory of this being "extensible with
new devices" was not consistent with practice.

In Ren-C, the shift has been gradually to providing "efficient-enough"
services which can be mixed with pure C, at either a %sys-core.h or
%rebol.h higher level API.

This takes an initial step toward tearing out the baked-in C tables
by moving the "device initialization" and "device shutdown" into
being functions called in extensions.  Those functions may be native,
usermode, or implemented with a mixture of both.