Clarify and improve some console isotope handling
The console responses to isotopes were (necessarily) some of the first
code written when the idea was being framed.  There were several
intermediary concepts that were eventually dropped (such as one where
a single tic ' would evaluate to a null isotope instead of pure null,
and ~null~ would produce pure null...but sanity won out on that idea).

This removes some of the traces of the dead evolution, and improves
the comments.