Fix stray "END" outside of parse in %prot-http
There was an END which was presumably trying to be inside a parse
rule, but was actually outside of it.  This wasn't causing a problem
because all paths in the parse would THROW.

(Presumably it was added somewhat robotically during the time when
PARSE did not enforce reaching END.)

This changes to use UPARSE for the task, which avoids the need for
the CATCH/THROW, making the code more succinct by actually using the
synthesized value.  Makes use of the newly-resurrected RETURN combinator
in order to return "abrupbtly".

(Also switches characters in the match to text strings, as the extra
# character makes it look junkier for what will likely be little
performance advantage.  And this is being done once per network request
so it's a silly place to even be looking for such optimization.)