Remove "View" from System Object, capture symbols
Most aspects of the "View" objects have been removed from the core,
with the idea that any GUI would be implemented in an extension.
The exception was that because EVENT! is based on a compression
scheme of putting only a finite number of "symbols" into a cell, those
symbol numbers have to be standardized by the core to implement
the EVENT!.

(This is a pretty ancillary concern, but it was kept in mind while
trying to implement EVENT! as an extension.)

There's a better way to keep that list of symbols around, which is
now done by symbols.r in the boot.  There's no need for the object
in the core.