More NULL fields in system.options, null from WHAT-DIR
This makes WHAT-DIR return null instead of BLANK! when it does not
know what directory it's in.  The advantage is that you will get
some noisiness if you try to say something like `append block what-dir`
instead of a no-op.

For similar reasons, other fields are changed in system.options.
The `suppress` and `loaded` fields are left as is, as an experiment.