Make boot banner show (none) for NULL fields
Previously this would show an underscore for BLANK!.  But changing the
fields to NULL made them more ornery, such that FORM would raise an
error.

In historical Rebol this would have said the word "none" as the raw
rendering of none.  This defaults the field to put nulls in parentheses
as (none).