Remove RDM_NULL flag
This was a flag related somehow to saying that a device was in a mode
where it would be acting like "/dev/null" and not have any output.  (?)

The only device that seemed to pretend to heed this was stdio, and that
was also seemingly spotty.  In any case, it feels like something that
should be addressed at a higher level in a way that is cross-cutting
and all "devices" or "ports" get it for free, vs. needing to check
every branch of every operation to see if they've been set in
"null mode".  :-/