Make StdIO no longer depend on the Rebol device model
The concept of the Rebol Device model was to provide a layer of
abstraction for modeling platform services.  It was agnostic to
Rebol (e.g. the APIs were defined in terms of pure C structs and
not REBVAL*) so higher level abstractions (such as ports) needed
to unpack and pack the data out of Rebol types to communicate with
this lower layer.

This layer was also entangled with the idea of eventing, and having
things like a network read or connection request receive notifications
when they were ready.

In the case of standard input and output, there was no asynchronous
processing being done...and there were only disadvantages to making
a layer that was not aware of Rebol values.  It has been an ongoing
process to try and move away from the device layer, and this takes
the long awaited step of actually deleting the Dev_StdIO "device"
entirely.

This means that the console port can just call functions that
interact with the appropriate platform APIs directly, which makes
it much easier to understand as well as more efficient.