Mitigate situation of buffered characters at program start
When starting the program, there is a period of time before the
console extension has been initialize and taken over control of the
"smart console"...doing the management of the cursor keys and echo.
During this startup period, any keys the user types will be echoed
to the screen.  This can cause confusion when the buffered characters
reach the console implementation and echos them again--because it
thinks they were not displayed.

Due to the lack of a general solution to the problem, this commit
prints a message to indicate it's out of sync and then goes to
a new line with the (potentially adjutsed) input.