Add new /HEAD and /TAIL DELIMIT feature, TEXT! only
DELIMIT was good for spacing, but in cases like inserting newlines one
often wants a newline at the tail as well as between the values.  This
points out that the /HEAD and /TAIL condition are special, so the
refinements are added.

A goal of DELIMIT is to be able to work on either blocks or TEXT! and
pass through the input as-is if it doesn't need delimiting.  The
/HEAD and /TAIL options still would apply to this case.  So this adds
that behavior, and also corrects that if the input had been an ISSUE!
the output would try to COPY that issue and not turn it to TEXT!.

This also consolidates the function that implemented DELIMIT (called
"Form_Reduce_Throws()") into DELIMIT itself.  Increasingly the design
is moving away from the idea of trying to optimize reusable bits of
code via the C and instead doing the reuse through the evaluator and
Rebol types.  This is particularly important in moving to a stackless
model, as any pure C code that tries to do evaluations would have to
maintain state and jump to the state locations--something that only
painstaking native authors will want to do.