Adjust HALT handling in Console
It's nice when the console's [You Halted] message doesn't appear
at random on a line that's in mid-printing, halfway through the screen.
This makes it print on its own line.

(Review making it a special color so it's distinguished from the rest
of the output.)

Also fixes an issue with the new ability to do a Ctrl-C while typing
input.  This avoids an unskinning response in that case.