Move READ-LINE to be a native in stdio extension
This is a bunch of interrelated changes related to getting the
READ-LINE native separated into its own function that can actually
read the standard input byte-by-byte in order to implement a
read-by-line facility.

The concept of a higher level layer which would read chunks of data
might seem interesting...but, there is generally significant buffering
in the I/O layer already.  Implementing one's own buffering on top of
it may actually slow things down.

As best possible, this commit is trying to nail down handling of
Ctrl-C as well as redirection of stdin and stdout.  However, those
facets are now going to be tested by GitHub Actions in order to
validate them across platforms and builds, and test more edge cases.
So this commit is being pushed to get the binaries to run those tests.