Relax the no-include-stdio.h rule
It's an uphill battle with modern compilers to cause errors by the
mere inclusion of stdio.h, as including one standard header can
pull in many of them--and it's becoming endemic that stdio.h is
included.

Attacking the problem will thus move to the linker phase, where
dependency analysis is better done in this day and age.