Add READ-CHAR For Reading One STDIO Character at a Time
This gives a facility for reading one character at a time from either
a smart terminal or a redirected input device.

It is truly a *character* based interface, so it is interpreting bytes
as UTF-8.  A separate READ-BYTE interface will be needed for reading
one byte at a time from files.

A speculative /VIRTUAL option is given to READ-CHAR that lets it
interpret cursor keys or ESCAPE, by returning them as WORD!.  This
may be changed to a different interface, something like READ-KEY