All extensions use STARTUP*/SHUTDOWN*, auto startup*
Prior to extensions having an ability to shutdown, many extensions had
shutdown functions that were written and just never got called.

Also, several had init functions called manually in the body of the
extension's module initialization.

This transitions to automatically injecting a call to a STARTUP*
function if one is present in the natives list, and it also standardizes
all the extensions to use the STARTUP* and SHUTDOWN* functions...meaning
many of the extensions are getting shutdown code called for the
first time.