Eliminate `error` field in PORT! object
R3-Alpha's original device model had a confusing integer-based error
encoding scheme that was separate from whatever error concept that
the language as a whole made (due to trying to make an abstracted OS
that wasn't necessarily for running Rebol).  This error scheme was
removed little by little until it was completely gone, consolidating
on the ERROR! and FAIL mechanics used system-wide.

An artifact of the old error mechanics was residually in the standard
definition for a PORT!.  This removes it, and also makes the
definition ues NULL instead of BLANK! for its fields by default.
(~unset~ would probably be an even better choice, but this is
just making a small step).