Remove window handle code and Reap_Process from Events
In the interest of continuing to simplify (and remove) the Reb_Device
code, this gets rid of some commented out code for event processing
related to creating an invisible window handle.  It also removes
Reap_Process(), which was added after the R3-Alpha open sourcing
by Atronix for some purpose that is currently not known.