Remove Reb_Device code from file and directory ports
The file and directory ports only provided synchronous I/O.  This
essentially meant that being cast in terms of the Rebol Device model
was nothing but busywork--creating lossy transformations between
Rebol types and fundamental C types.

This removes the device code, instead making the device commands into
functions that are defined either for Windows or POSIX.

This is a first-cut at the transition, and attempts to make no
changes to the logic.  It changes from using a REBREQ to a "FILEREQ"
structure, and moves a few fields from that out into parameters
when it is not appropriate to keep as port state.